SVAI Hackathon/ai-cofounder-mvp/backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
crewai==0.28.8
boto3==1.35.32
snowflake-connector-python==3.12.2
yfinance==0.2.44
httpx==0.27.2
